finish minor refactoring
