+++ pull_request_template.md
@@ -1,0 +1,18 @@
+### Description
+_TLDR on the summary of the changes, for what and why. Also list any dependencies that are required for this change._
+
+### Type of change
+_Please add 'x' for the type of this change._ 
+- [ ] Feature
+- [ ] Bug Fix
+- [ ] Chore
+
+### Demo 
+_Add URL for video recordings._
+
+
+### Checklist
+_Please add 'x' on the checklist items to confirm._
+- [ ] Reviewed and cleaned my code
+- [ ] Added helpful comments on my code
+- [ ] Tested in vercel deployment
