+++ src/components/Modal.tsx
@@ -1,41 +1,40 @@
 'use client';
 
-import React, { ChangeEvent, useState } from 'react';
-import { Modal, ToggleSwitch } from 'flowbite-react';
+import React from 'react';
+import { Modal } from 'flowbite-react';
 import { Button } from './Button';
 
-interface ModalProps {
+type RenderElement<T> = {
+    value: any;
+    element: T;
+};
+
+type ModalFields<T> = {
+    title: string;
+    key: string;
+    render?: (element: RenderElement<T>) => JSX.Element | string | null;
+};
+
+interface ModalProps<T> {
     openModal: boolean;
     onClose: () => void;
     onClick?: (e: any) => void;
     title?: string;
     haveBody?: boolean;
-    buttonAction?: string;
-    handleTextChange?: (e: ChangeEvent<HTMLTextAreaElement>) => void;
-    handleCollegeChange?: (e: ChangeEvent<HTMLSelectElement>) => void;
-    handleAuthorChange?: (e: ChangeEvent<HTMLInputElement>) => void;
-    onSubmitPost?: () => void;
+    modalFields: ModalFields<T>[];
     text?: string;
-    maxLength?: number;
-    isAnon?: boolean;
-    setIsAnon?: (value: boolean) => void;
+    buttonAction?: string;
 }
 
-export const ModalComponent: React.FC<ModalProps> = ({
+export const ModalComponent: React.FC<ModalProps<any>> = ({
     openModal,
     onClose,
     onClick,
     title,
     haveBody = true,
-    buttonAction,
-    handleTextChange,
-    handleCollegeChange,
-    handleAuthorChange,
-    onSubmitPost,
+    modalFields,
     text,
-    maxLength,
-    isAnon,
-    setIsAnon,
+    buttonAction,
 }) => {
     return (
         <Modal dismissible show={openModal} onClose={onClose}>
@@ -46,13 +45,23 @@ export const ModalComponent: React.FC<ModalProps> = ({
             {haveBody && (
                 <Modal.Body>
                     <div className="space-y-6">
-                        
+                        <p className="text-base leading-relaxed text-gray-500">
+                            {text}
+                        </p>
+
+                        {modalFields?.map((field: any) => (
+                            <div key={field.key}>
+                                {field.render
+                                ? field.render({ value: field, element: field })
+                                : null}
+                            </div>
+                        ))}
                     </div>
                 </Modal.Body>
             )}
 
             {/* modal footer */}
-            <Modal.Footer className="flex items-center">
+            <Modal.Footer className="flex items-center justify-end">
                 {/* actions */}
                 <Button label="Cancel" onClick={onClose} danger />
                 <Button label={buttonAction} onClick={onClick} action />
